Chop into multiple files
