+++ src/infrastructure/prun-ui/i18n/localization.gen.d.ts
