+++ src/infrastructure/prun-ui/i18n/materials.ts
@@ -1,0 +1,1 @@
+import { L } from './index';
